backend/requirements.txt
Django==3.2
djangorestframework
djoser